requirements.txt
requests>=2.32.3
black>=24.10.0
schema>=0.7.7
PyYAML>=6.0.2
